+++ iui/examples/basic.rs
@@ -1,13 +1,13 @@
 extern crate iui;
+use iui::controls::{Button, Group, Label, VerticalBox};
 use iui::prelude::*;
-use iui::controls::{Label, Button, VerticalBox, Group};
 
 fn main() {
     // Initialize the UI library
     let ui = UI::init().expect("Couldn't initialize UI library");
     // Create a window into which controls can be placed
     let mut win = Window::new(&ui, "Test App", 200, 200, WindowType::NoMenubar);
-    
+
     // Create a vertical layout to hold the controls
     let mut vbox = VerticalBox::new(&ui);
     vbox.set_padded(&ui, true);
@@ -39,11 +39,11 @@ fn main() {
     label_text.push_str("🎉 用户界面 사용자 인터페이스");
     let label = Label::new(&ui, &label_text);
 
-    vbox.append(&ui, label, LayoutStrategy::Stretchy);
-    group_vbox.append(&ui, button, LayoutStrategy::Compact);
-    group_vbox.append(&ui, quit_button, LayoutStrategy::Compact);
+    vbox.add_stretchy(&ui, label);
+    group_vbox.add(&ui, button);
+    group_vbox.add(&ui, quit_button);
     group.set_child(&ui, group_vbox);
-    vbox.append(&ui, group, LayoutStrategy::Compact);
+    vbox.add(&ui, group);
 
     // Actually put the button in the window
     win.set_child(&ui, vbox);