+++ iui/examples/files.rs
@@ -2,11 +2,11 @@
 //! and the Window::modal_err() call to display modal dialog boxes.
 
 extern crate iui;
+use iui::controls::{Button, MultilineEntry, VerticalBox};
 use iui::prelude::*;
-use iui::controls::{VerticalBox, MultilineEntry, Button};
-use std::io::prelude::*;
 use std::error::Error;
 use std::fs::File;
+use std::io::prelude::*;
 
 fn main() {
     // Initialize the UI
@@ -19,8 +19,8 @@ fn main() {
     // Set up the application's layout
     let mut window = Window::new(&ui, "Save Buffer to File", 640, 480, WindowType::NoMenubar);
     let mut vbox = VerticalBox::new(&ui);
-    vbox.append(&ui, entry.clone(), LayoutStrategy::Stretchy);
-    vbox.append(&ui, button.clone(), LayoutStrategy::Compact);
+    vbox.add_stretchy(&ui, entry.clone());
+    vbox.add(&ui, button.clone());
     window.set_child(&ui, vbox);
     window.show(&ui);
 
@@ -32,17 +32,38 @@ fn main() {
         move |_| {
             if let Some(path) = window.save_file(&ui) {
                 let mut file = match File::create(&path) {
-                    Err(why) => { window.modal_err(&ui, "I/O Error", &format!("Could not open file {}: {}", path.display(), why.description())); return; }
-                    Ok(f) => f
+                    Err(why) => {
+                        window.modal_err(
+                            &ui,
+                            "I/O Error",
+                            &format!(
+                                "Could not open file {}: {}",
+                                path.display(),
+                                why.description()
+                            ),
+                        );
+                        return;
+                    }
+                    Ok(f) => f,
                 };
                 match file.write_all(entry.value(&ui).as_bytes()) {
-                    Err(why) => { window.modal_err(&ui, "I/O Error", &format!("Could not write to file {}: {}", path.display(), why.description())); return; }
-                    Ok(_) => ()
+                    Err(why) => {
+                        window.modal_err(
+                            &ui,
+                            "I/O Error",
+                            &format!(
+                                "Could not write to file {}: {}",
+                                path.display(),
+                                why.description()
+                            ),
+                        );
+                        return;
+                    }
+                    Ok(_) => (),
                 };
-            }    
+            }
         }
     });
 
     ui.main();
-
 }