+++ iui/examples/inputs.rs
@@ -1,10 +1,13 @@
 //! Demonstrates a mutable application state manipulated over a number of UIs
 
 extern crate iui;
+use iui::controls::{
+    Entry, Group, HorizontalBox, HorizontalSeparator, Label, MultilineEntry, Slider, Spacer,
+    Spinbox, VerticalBox,
+};
 use iui::prelude::*;
-use iui::controls::{Label, Spinbox, Slider, Entry, MultilineEntry, VerticalBox, HorizontalBox, HorizontalSeparator, Group, Spacer};
-use std::rc::Rc;
 use std::cell::RefCell;
+use std::rc::Rc;
 
 /// This struct will hold the values that multiple callbacks will need to access.
 struct State {
@@ -19,7 +22,12 @@ fn main() {
     let ui = UI::init().unwrap();
 
     // Initialize the state of the application.
-    let state = Rc::new(RefCell::new(State { slider_val: 0, spinner_val: 0, entry_val: "".into(), multi_val: "".into() }));
+    let state = Rc::new(RefCell::new(State {
+        slider_val: 0,
+        spinner_val: 0,
+        entry_val: "".into(),
+        multi_val: "".into(),
+    }));
 
     // Set up the inputs for the application.
     // While it's not necessary to create a block for this, it makes the code a lot easier
@@ -39,13 +47,13 @@ fn main() {
         // Add everything in hierarchy
         // Note the reverse order here. Again, it's not necessary, but it improves
         // readability.
-        input_vbox.append(&ui, slider.clone(), LayoutStrategy::Compact);
-        input_vbox.append(&ui, spinner.clone(), LayoutStrategy::Compact);
-        input_vbox.append(&ui, Spacer::new(&ui), LayoutStrategy::Compact);
-        input_vbox.append(&ui, HorizontalSeparator::new(&ui), LayoutStrategy::Compact);
-        input_vbox.append(&ui, Spacer::new(&ui), LayoutStrategy::Compact);
-        input_vbox.append(&ui, entry.clone(), LayoutStrategy::Compact);
-        input_vbox.append(&ui, multi.clone(), LayoutStrategy::Stretchy);
+        input_vbox.add(&ui, slider.clone());
+        input_vbox.add(&ui, spinner.clone());
+        input_vbox.add(&ui, Spacer::new(&ui));
+        input_vbox.add(&ui, HorizontalSeparator::new(&ui));
+        input_vbox.add(&ui, Spacer::new(&ui));
+        input_vbox.add(&ui, entry.clone());
+        input_vbox.add_stretchy(&ui, multi.clone());
         input_group.set_child(&ui, input_vbox);
         (input_group, slider, spinner, entry, multi)
     };
@@ -59,18 +67,24 @@ fn main() {
         let sub_label = Label::new(&ui, "");
         let text_label = Label::new(&ui, "");
         let bigtext_label = Label::new(&ui, "");
-        output_vbox.append(&ui, add_label.clone(), LayoutStrategy::Compact);
-        output_vbox.append(&ui, sub_label.clone(), LayoutStrategy::Compact);
-        output_vbox.append(&ui, text_label.clone(), LayoutStrategy::Compact);
-        output_vbox.append(&ui, bigtext_label.clone(), LayoutStrategy::Stretchy);
+        output_vbox.add(&ui, add_label.clone());
+        output_vbox.add(&ui, sub_label.clone());
+        output_vbox.add(&ui, text_label.clone());
+        output_vbox.add_stretchy(&ui, bigtext_label.clone());
         output_group.set_child(&ui, output_vbox);
-        (output_group, add_label, sub_label, text_label, bigtext_label)
+        (
+            output_group,
+            add_label,
+            sub_label,
+            text_label,
+            bigtext_label,
+        )
     };
 
     // This horizontal box will arrange the two groups of controls.
     let mut hbox = HorizontalBox::new(&ui);
-    hbox.append(&ui, input_group, LayoutStrategy::Stretchy);
-    hbox.append(&ui, output_group, LayoutStrategy::Stretchy);
+    hbox.add_stretchy(&ui, input_group);
+    hbox.add_stretchy(&ui, output_group);
 
     // The window allows all constituent components to be displayed.
     let mut window = Window::new(&ui, "Input Output Test", 300, 150, WindowType::NoMenubar);
@@ -82,24 +96,31 @@ fn main() {
 
     slider.on_changed(&ui, {
         let state = state.clone();
-        move |val| { state.borrow_mut().slider_val = val; }
+        move |val| {
+            state.borrow_mut().slider_val = val;
+        }
     });
 
     spinner.on_changed(&ui, {
         let state = state.clone();
-        move |val| { state.borrow_mut().spinner_val = val; }
+        move |val| {
+            state.borrow_mut().spinner_val = val;
+        }
     });
 
     entry.on_changed(&ui, {
         let state = state.clone();
-        move |val| { state.borrow_mut().entry_val = val; }
+        move |val| {
+            state.borrow_mut().entry_val = val;
+        }
     });
 
     multi.on_changed(&ui, {
         let state = state.clone();
-        move |val| { state.borrow_mut().multi_val = val; }
+        move |val| {
+            state.borrow_mut().multi_val = val;
+        }
     });
-
 
     // Rather than just invoking ui.run(), using EventLoop gives a lot more control
     // over the user interface event loop.
@@ -115,8 +136,14 @@ fn main() {
             let state = state.borrow();
 
             // Update all the labels
-            add_label.set_text(&ui, &format!("Added: {}", state.slider_val + state.spinner_val));
-            sub_label.set_text(&ui, &format!("Subtracted: {}", state.slider_val - state.spinner_val));
+            add_label.set_text(
+                &ui,
+                &format!("Added: {}", state.slider_val + state.spinner_val),
+            );
+            sub_label.set_text(
+                &ui,
+                &format!("Subtracted: {}", state.slider_val - state.spinner_val),
+            );
             text_label.set_text(&ui, &format!("Text: {}", state.entry_val));
             bigtext_label.set_text(&ui, &format!("Multiline Text: {}", state.multi_val));
         }