+++ iui/src/lib.rs
@@ -9,22 +9,22 @@
 //! perform precisely the same on all platforms. These inconsistencies are marked.
 //!
 //! To use the library, add the following to your `Cargo.toml`:
-//! 
+//!
 //! ```toml
 //! "iui" = "0.3"
 //! ```
 //!
 //! Most of the functionality of the crate is exposed via the [UI](struct.UI.html) RAII guard, which handles all initialization and cleanup for the
 //! underlying library.
-//! 
-//! After initialization, all the functionality used for creating actual UIs is in the [`controls`](controls/index.html) module. 
-//! 
+//!
+//! After initialization, all the functionality used for creating actual UIs is in the [`controls`](controls/index.html) module.
+//!
 //! Fine-grained control of the event loop is avilable via the [`EventLoop`](struct.EventLoop.html) struct.
 //! Be aware the Cocoa (GUI toolkit on Mac OS) requires that the _first thread spawned_ controls
 //! the UI, so do _not_ spin off your UI interactions into an alternative thread. You're likely to
 //! have problems on Mac OS.
 //!
-//! For code examples, see the [examples](https://github.com/LeoTindall/libui-rs/blob/master/iui/examples/) 
+//! For code examples, see the [examples](https://github.com/LeoTindall/libui-rs/blob/master/iui/examples/)
 //! directory.
 
 #[macro_use]